+++ src/main/kotlin/com/nylas/models/MessageFields.kt
@@ -8,4 +8,10 @@ enum class MessageFields {
 
   @Json(name = "include_headers")
   INCLUDE_HEADERS,
+
+  @Json(name = "include_tracking_options")
+  INCLUDE_TRACKING_OPTIONS,
+
+  @Json(name = "raw_mime")
+  RAW_MIME,
 }